+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+<a name="6.0.7"></a>
+## 6.0.7 (2016-07-14)
+
+
+### Bug Fixes
+
+* **forms:** safari/phantom don&#x27;t have setImmediate ([b043f17](https://github.com/pivotal-cf/pivotal-ui/commit/b043f17))
+
+
+
+<a name="6.0.6"></a>
+## 6.0.6 (2016-07-14)
+
+
+### Bug Fixes
+
+* **forms:** Validator now works w/ input-error evt ([3257d26](https://github.com/pivotal-cf/pivotal-ui/commit/3257d26))
+
+
+
 <a name="6.0.6"></a>
 ## 6.0.6 (2016-07-14)